moved files to configs directory
