+++ bin/input-remapper-control
@@ -30,7 +30,7 @@ import logging
 import subprocess
 
 from inputremapper.logger import logger, update_verbosity, log_info, add_filehandler
-from inputremapper.migrations import migrate
+from inputremapper.configs.migrations import migrate
 from inputremapper.configs.global_config import global_config
 
 # import inputremapper modules as late as possible to make sure the correct
@@ -79,7 +79,7 @@ def utils(options):
             print(group.key)
 
     if options.key_names:
-        from inputremapper.system_mapping import system_mapping
+        from inputremapper.configs.system_mapping import system_mapping
         print('\n'.join(system_mapping.list_names()))
 
 
@@ -88,7 +88,7 @@ def communicate(options, daemon):
     # import stuff late to make sure the correct log level is applied
     # before anything is logged
     from inputremapper.groups import groups
-    from inputremapper.paths import USER
+    from inputremapper.configs.paths import USER
 
     def require_group():
         if options.device is None:
@@ -216,7 +216,7 @@ def main(options):
 
     logger.debug('Call for "%s"', sys.argv)
 
-    from inputremapper.paths import USER
+    from inputremapper.configs.paths import USER
     boot_finished = systemd_finished()
     is_root = USER == "root"
     is_autoload = options.command == AUTOLOAD
