+++ inputremapper/daemon.py
@@ -39,11 +39,11 @@ from gi.repository import GLib
 
 from inputremapper.logger import logger, is_debug
 from inputremapper.injection.injector import Injector, UNKNOWN
-from inputremapper.preset import Preset
+from inputremapper.configs.preset import Preset
 from inputremapper.configs.global_config import global_config
-from inputremapper.system_mapping import system_mapping
+from inputremapper.configs.system_mapping import system_mapping
 from inputremapper.groups import groups
-from inputremapper.paths import get_config_path, USER
+from inputremapper.configs.paths import get_config_path, USER
 from inputremapper.injection.macros.macro import macro_variables
 from inputremapper.injection.global_uinputs import global_uinputs
