+++ inputremapper/injection/context.py
@@ -23,7 +23,7 @@
 
 from inputremapper.logger import logger
 from inputremapper.injection.macros.parse import parse, is_this_a_macro
-from inputremapper.system_mapping import system_mapping
+from inputremapper.configs.system_mapping import system_mapping
 from inputremapper.configs.global_config import NONE, MOUSE, WHEEL, BUTTONS
