+++ inputremapper/ipc/pipe.py
@@ -42,7 +42,7 @@ import time
 import json
 
 from inputremapper.logger import logger
-from inputremapper.paths import mkdir, chown
+from inputremapper.configs.paths import mkdir, chown
 
 
 class Pipe:
