+++ tests/unit/test_event_producer.py
@@ -36,7 +36,7 @@ from evdev.ecodes import (
 )
 
 from inputremapper.configs.global_config import global_config
-from inputremapper.preset import Preset
+from inputremapper.configs.preset import Preset
 from inputremapper.injection.context import Context
 from inputremapper.injection.mapping_handlers.joystick_to_mouse import (
     JoystickToMouse,
