+++ tests/unit/test_migrations.py
@@ -20,10 +20,10 @@ import json
 
 from evdev.ecodes import EV_KEY, EV_ABS, ABS_HAT0X
 
-from inputremapper.migrations import migrate, config_version
-from inputremapper.preset import Preset
+from inputremapper.configs.migrations import migrate, config_version
+from inputremapper.configs.preset import Preset
 from inputremapper.configs.global_config import global_config
-from inputremapper.paths import touch, CONFIG_PATH, mkdir, get_preset_path
+from inputremapper.configs.paths import touch, CONFIG_PATH, mkdir, get_preset_path
 from inputremapper.key import Key
 from inputremapper.user import HOME
